Refaktorering: Innførte Fødselsnummer-klasse
