+++ src/main/kotlin/no/nav/medlemskap/regler/funksjoner/RelasjonFunksjoner.kt
@@ -3,6 +3,7 @@ package no.nav.medlemskap.regler.funksjoner
 import no.bekk.bekkopen.person.FodselsnummerValidator
 import no.nav.medlemskap.domene.Familierelasjon
 import no.nav.medlemskap.domene.Familierelasjonsrolle
+import no.nav.medlemskap.domene.Fødselsnummer.Companion.hentBursdagsAar
 import no.nav.medlemskap.domene.Personhistorikk
 import no.nav.medlemskap.domene.Sivilstandstype
 import java.util.*
@@ -11,27 +12,6 @@ object RelasjonFunksjoner {
 
     fun String.erBarnUnder25Aar() =
         (Calendar.getInstance().get(Calendar.YEAR) - this.hentBursdagsAar().toInt()) <= 25
-
-    fun String.hentBursdagsAar(): String {
-        return this.hentAarHundre() + this.hent2DigitBursdagsAar()
-    }
-
-    fun String.hentAarHundre(): String? {
-        val individnummer: Int = getIndividnummer().toInt()
-        val birthYear: Int = hent2DigitBursdagsAar().toInt()
-
-        return when {
-            individnummer <= 499 -> "19"
-            individnummer >= 500 && birthYear < 40 -> "20"
-            individnummer in 500..749 && birthYear >= 54 -> "18"
-            individnummer >= 900 && birthYear > 39 -> "19"
-            else -> null
-        }
-    }
-
-    fun String.hent2DigitBursdagsAar() = this.substring(4, 6)
-
-    fun String.getIndividnummer() = this.substring(6, 9)
 
     fun hentFnrTilEktefelle(personHistorikkFraPdl: Personhistorikk?): String? {
         val fnrTilEktefelle =
